requirements.txt
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
mss>=9.0.0
imagehash>=4.3.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
scikit-learn>=1.3.0
